+++ attendance-ios/Source/SignUp/SignUpPositionViewController.swift
@@ -48,19 +48,10 @@ final class SignUpPositionViewController: UIViewController {
         return button
     }()
 
-    private let backButton: UIButton = {
-        let button = UIButton()
-        button.setImage(UIImage(named: "back"), for: .normal)
-        return button
-    }()
-
     private let alertView: AlertView = {
         let view = AlertView()
         view.isHidden = true
-        view.configureUI(text: "입력을 취소할까요?",
-                         subText: "언제든 다시 돌아올 수 있어요",
-                         leftButtonText: "아니요",
-                         rightButtonText: "취소합니다")
+        view.configureUI(text: "입력을 취소할까요?", subText: "언제든 다시 돌아올 수 있어요", leftButtonText: "아니요", rightButtonText: "취소합니다")
         return view
     }()
 
@@ -83,13 +74,15 @@ final class SignUpPositionViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        bindViewModel()
         bindButton()
+
         setupDelegate()
         setupCollectionView()
+
         configureUI()
         configureLayout()
         configureAlertViewLayout()
+        addNavigationBackButton()
     }
 
     override func viewWillAppear(_ animated: Bool) {
@@ -97,26 +90,20 @@ final class SignUpPositionViewController: UIViewController {
         navigationItem.hidesBackButton = true
     }
 
+    override func navigationBackButtonTapped() {
+        alertView.isHidden.toggle()
+    }
+
 }
 
 // MARK: - Bind
 private extension SignUpPositionViewController {
-
-    func bindViewModel() {
-
-    }
 
     func bindButton() {
         nextButton.rx.controlEvent([.touchUpInside])
             .asObservable()
             .subscribe(onNext: { [weak self] _ in
                 self?.goToTeamVC()
-            }).disposed(by: disposeBag)
-
-        backButton.rx.controlEvent([.touchUpInside])
-            .asObservable()
-            .subscribe(onNext: { [weak self] _ in
-                self?.alertView.isHidden.toggle()
             }).disposed(by: disposeBag)
 
         alertView.rightButton.rx.controlEvent([.touchUpInside])
@@ -219,13 +206,8 @@ private extension SignUpPositionViewController {
     }
 
     func configureLayout() {
-        view.addSubviews([backButton, titleLabel, collectionView, nextButton])
+        view.addSubviews([titleLabel, collectionView, nextButton])
 
-        backButton.snp.makeConstraints {
-            $0.top.equalToSuperview().offset(56)
-            $0.left.equalToSuperview().offset(Constants.padding/2)
-            $0.width.height.equalTo(40)
-        }
         titleLabel.snp.makeConstraints {
             $0.top.equalToSuperview().offset(120)
             $0.left.right.equalToSuperview().inset(Constants.padding)
