+++ attendance-ios/Source/SignUp/SignUpTeamViewController.swift
@@ -71,19 +71,10 @@ final class SignUpTeamViewController: UIViewController {
         return button
     }()
 
-    private let backButton: UIButton = {
-        let button = UIButton()
-        button.setImage(UIImage(named: "back"), for: .normal)
-        return button
-    }()
-
     private let alertView: AlertView = {
         let view = AlertView()
         view.isHidden = true
-        view.configureUI(text: "입력을 취소할까요?",
-                         subText: "언제든 다시 돌아올 수 있어요",
-                         leftButtonText: "아니요",
-                         rightButtonText: "취소합니다")
+        view.configureUI(text: "입력을 취소할까요?", subText: "언제든 다시 돌아올 수 있어요", leftButtonText: "아니요", rightButtonText: "취소합니다")
         return view
     }()
 
@@ -115,11 +106,16 @@ final class SignUpTeamViewController: UIViewController {
         configureUI()
         configureLayout()
         configureAlertViewLayout()
+        addNavigationBackButton()
     }
 
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
         navigationItem.hidesBackButton = true
+    }
+
+    override func navigationBackButtonTapped() {
+        alertView.isHidden.toggle()
     }
 
 }
@@ -132,12 +128,6 @@ private extension SignUpTeamViewController {
             .asObservable()
             .subscribe(onNext: { [weak self] _ in
                 self?.viewModel.input.registerInfo.accept(())
-            }).disposed(by: disposeBag)
-
-        backButton.rx.controlEvent([.touchUpInside])
-            .asObservable()
-            .subscribe(onNext: { [weak self] _ in
-                self?.alertView.isHidden.toggle()
             }).disposed(by: disposeBag)
 
         alertView.rightButton.rx.controlEvent([.touchUpInside])
@@ -191,12 +181,6 @@ private extension SignUpTeamViewController {
             .asObservable()
             .subscribe(onNext: { [weak self] _ in
                 self?.viewModel.registerInfo()
-            }).disposed(by: disposeBag)
-
-        backButton.rx.controlEvent([.touchUpInside])
-            .asObservable()
-            .subscribe(onNext: { [weak self] _ in
-                self?.alertView.isHidden.toggle()
             }).disposed(by: disposeBag)
 
         alertView.rightButton.rx.controlEvent([.touchUpInside])
@@ -331,13 +315,8 @@ private extension SignUpTeamViewController {
     }
 
     func configureLayout() {
-        view.addSubviews([backButton, titleLabel, teamTypeCollectionView, subTitleLabel, teamNumberCollectionView, okButton])
+        view.addSubviews([titleLabel, teamTypeCollectionView, subTitleLabel, teamNumberCollectionView, okButton])
 
-        backButton.snp.makeConstraints {
-            $0.top.equalToSuperview().offset(56)
-            $0.left.equalToSuperview().offset(Constants.padding/2)
-            $0.width.height.equalTo(40)
-        }
         titleLabel.snp.makeConstraints {
             $0.top.equalToSuperview().offset(120)
             $0.left.right.equalToSuperview().inset(Constants.padding)
